requirements.txt
requests==2.28.1
dnspython==2.2.1
aiostun==0.2.0